solved

+++ binary_search/BOJ1654.c
@@ -6,7 +6,7 @@ typedef long long ll;
 
 int check(int *arr, int n, int mid, int t);
 int max(int a, int b);
-int _maxLen(int *arr, int n, int s, int e, int t);
+int _maxLen(int *arr, int n, ll s, int e, int t);
 int maxLen(int *arr, int n, int t);
 
 int main() {
@@ -36,16 +36,16 @@ int max(int a, int b) {
     return a > b ? a : b;
 }
 
-int _maxLen(int *arr, int n, int s, int e, int t) {
+int _maxLen(int *arr, int n, ll s, int e, int t) {
     if (s > e)
         return 0;
     
     int mid = s + (e - s + 1) / 2;
 
     if (check(arr, n, mid, t))
-        return max(_maxLen(arr, n, mid + 1, e, t), mid);
+        return max(_maxLen(arr, n, (ll) mid + 1, e, t), mid);
     else 
-        return _maxLen(arr, n, s, mid - 1, t);
+        return _maxLen(arr, n, (ll) s, mid - 1, t);
 }
 
 int maxLen(int *arr, int n, int t) {
